Fix some read bugs and add a simple read test (#33)

+++ cpp/src/storage/options.h
@@ -20,7 +20,7 @@ struct ReadOptions {
   FilterSet filters;
   std::vector<std::string> columns;  // must have pk and version
   int limit = -1;
-  int version = -1;
+  int64_t version = -1;
 
   static ReadOptions& default_read_options() {
     static ReadOptions options;
